MDL-51576 navigation: stop course nav duplication
The course navigation was getting duplicate nodes because
it was loading the nav tree plugin functions twice. Once
specifically for the reports and then a second time more generally.

The code will now remember which plugin function have been loaded
and wont execute those plugin functions a second time, so we don't
end up with duplicate nodes in the course nav tree.