MDL-51739 theme: allow config to lock theme selector
If the theme property is set in config.php it will lock
the site to that config and not allow the admin to change
the theme using the theme selector.

It will also restrict the list of themes shown in the admin
nav tree to just the config defined themes, so that only those
themes' settings can be modified.