MDL-53311 navigation: Move treeitem role on to p
Previously the li elements had the treeitem role, which was not invalide
but it meant that when a menu item expanded, the entire li was clickable
and clicking anywhere that was not another menu item would cause the
(sub)tree to collapse. This patch moves the treeitem role on to the p
element (which contains the node name) and uses aria-owns to expand
and collapse the subtrees.