MDL-66477 core: only render region main settings if no settings block
We added a change to allow the region main settings menu to be
rendered in the header rather that the top of the body however
this also meant that the settings were shown on themes (e.g. Classic)
that didn't otherwise show the region main settings.

The simplest solution seems to be to not render the region main
settings in the header if the settings block is on the page.